Fix long long bug in validator, add st2 to should-fail in bkorecic-int.cpp

+++ baldosas/testplan/validators/st3.cpp
@@ -4,10 +4,10 @@ const long long UPPER = 1'000'000'000'000'000'000;
 
 int main() {
   registerValidation();
-  int n = inf.readLong(1LL, UPPER, "n");
+  long long n = inf.readLong(1LL, UPPER, "n");
   ensuref(n%2 == 0, "n debe ser par");
   inf.readSpace();
-  int m = inf.readLong(1LL, UPPER, "m");
+  long long m = inf.readLong(1LL, UPPER, "m");
   ensuref(m%2 == 0, "m debe ser par");
   ensuref(n <= UPPER/m, "n*m debe ser menor o igual a 1e18");
   inf.readEoln();
